feat: improve scripts for ES index creation/deletion
